feat: Add Mongoose model for Blog
Add the Mongoose schema and model definition for the Blog entity in the Blog.js file. The schema includes fields for badges, image, title, description, author, and date. The date field is set to the current date by default using the Date.now function. This model will be used for interacting with the MongoDB database.